Added delete user functionality in views/user.py
The delete functionality has a boilerplate template, but we won't use it
for now. We have too many database tables depending on the user id as a
foreign key. So, if we expose the delete user functionality now, each
time we change something, we have to change this function as well