Solves bug when creating a new form field.
PROBLEM: When the user added a field, clicked on
"save" in Form Edit, the new field was correctly
added, but if the user clicked on save again the
field was added other time. This happened because
we used field_id = 'new' when the field was not
added to the DB yet.

SOLUTION: Return on field update successful action
a JSON with the new field's ids, and we update
the JS object (fields_json) correctly. Next time
this field will be updated, not created again.