Close bug #205 Form names validation
When we switched to json, deform validation stopped being used.
Here I put deform validation back; however, in the future we may
have to create our own javascript for error visualization.

The Form Properties tab is now rendered after every save; if
the user presses save once and gets validation errors and then
solves then, the subsequent save causes the validation errors to
disappear visually.

The form ID is now passed as a variable to the manager, instead of
residing in a hidden field.

The form title is no longer required, but the 255 char limit is
respected.

The URL had been changed to /form/update, but now it is back to
/form/edit

All console.log() calls removed.