Hide field label instead of showing "Question"
When the user deletes the whole label of a field, instead of
"Question" appearing in the visualization, the label must
become blank. Also in the entry and in the report screens.

Closes #653.

To complete this I had to test every field, so I also had to make
the list field save again.